Initial set of tests are in place
Even though the tests don't really test for anything as they're not using mocks
or anything fancy, they do have one sanity test so that more tests can be added
later.
